+++ rpc/src/client/transport/websocket.rs
@@ -20,10 +20,25 @@ use async_tungstenite::WebSocketStream;
 use futures::{SinkExt, StreamExt};
 use serde::{Deserialize, Serialize};
 use std::borrow::Cow;
+use std::ops::Add;
 use std::str::FromStr;
 use tendermint::net;
 use tokio::net::TcpStream;
 use tokio::task::JoinHandle;
+use tokio::time::{Duration, Instant};
+
+// WebSocket connection times out if we haven't heard anything at all from the
+// server in this long.
+//
+// Taken from https://github.com/tendermint/tendermint/blob/309e29c245a01825fc9630103311fd04de99fa5e/rpc/jsonrpc/server/ws_handler.go#L27
+const RECV_TIMEOUT_SECONDS: u64 = 30;
+
+const RECV_TIMEOUT: Duration = Duration::from_secs(RECV_TIMEOUT_SECONDS);
+
+// How frequently to send ping messages to the WebSocket server.
+//
+// Taken from https://github.com/tendermint/tendermint/blob/309e29c245a01825fc9630103311fd04de99fa5e/rpc/jsonrpc/server/ws_handler.go#L28
+const PING_INTERVAL: Duration = Duration::from_secs((RECV_TIMEOUT_SECONDS * 9) / 10);
 
 /// Tendermint RPC client that provides [`Event`] subscription capabilities
 /// over JSON-RPC over a WebSocket connection.
@@ -34,6 +49,23 @@ use tokio::task::JoinHandle;
 /// This driver is spawned as the client is created.
 ///
 /// To terminate the client and the driver, simply use its [`close`] method.
+///
+/// ### Timeouts
+///
+/// The WebSocket client connection times out after 30 seconds if it does not
+/// receive anything at all from the server. This will automatically return
+/// errors to all active subscriptions and terminate them.
+///
+/// This is not configurable at present.
+///
+/// ### Keep-Alive
+///
+/// The WebSocket client implements a keep-alive mechanism whereby it sends a
+/// PING message to the server every 27 seconds, matching the PING cadence of
+/// the Tendermint server (see [this code](tendermint-websocket-ping) for
+/// details).
+///
+/// This is not configurable at present.
 ///
 /// ## Examples
 ///
@@ -75,6 +107,7 @@ use tokio::task::JoinHandle;
 ///
 /// [`Event`]: ./event/struct.Event.html
 /// [`close`]: struct.WebSocketClient.html#method.close
+/// [tendermint-websocket-ping]: https://github.com/tendermint/tendermint/blob/309e29c245a01825fc9630103311fd04de99fa5e/rpc/jsonrpc/server/ws_handler.go#L28
 #[derive(Debug)]
 pub struct WebSocketClient {
     host: String,
@@ -166,12 +199,16 @@ impl WebSocketSubscriptionDriver {
     }
 
     async fn run(mut self) -> Result<()> {
-        // TODO(thane): Should this loop initiate a keepalive (ping) to the
-        //              server on a regular basis?
+        let mut ping_interval =
+            tokio::time::interval_at(Instant::now().add(PING_INTERVAL), PING_INTERVAL);
+        let mut recv_timeout = tokio::time::delay_for(PING_INTERVAL);
         loop {
             tokio::select! {
                 Some(res) = self.stream.next() => match res {
-                    Ok(msg) => self.handle_incoming_msg(msg).await?,
+                    Ok(msg) => {
+                        recv_timeout.reset(Instant::now().add(PING_INTERVAL));
+                        self.handle_incoming_msg(msg).await?
+                    },
                     Err(e) => return Err(
                         Error::websocket_error(
                             format!("failed to read from WebSocket connection: {}", e),
@@ -192,6 +229,13 @@ impl WebSocketSubscriptionDriver {
                     } => self.unsubscribe(id, query, result_tx).await?,
                     SubscriptionDriverCmd::Terminate => return self.close().await,
                 },
+                _ = ping_interval.next() => self.ping().await?,
+                _ = &mut recv_timeout => {
+                    return Err(Error::websocket_error(format!(
+                        "reading from WebSocket connection timed out after {} seconds",
+                        RECV_TIMEOUT.as_secs()
+                    )));
+                }
             }
         }
     }
@@ -313,7 +357,6 @@ impl WebSocketSubscriptionDriver {
                         if let Some(event_tx) = self.router.get_active_subscription_mut(
                             &SubscriptionId::from_str(&req_id).unwrap(),
                         ) {
-                            // TODO(thane): Does an error here warrant terminating the subscription, or the driver?
                             let _ = event_tx.send(Err(e)).await;
                         }
                     }
@@ -324,6 +367,10 @@ impl WebSocketSubscriptionDriver {
 
     async fn pong(&mut self, v: Vec<u8>) -> Result<()> {
         self.send_msg(Message::Pong(v)).await
+    }
+
+    async fn ping(&mut self) -> Result<()> {
+        self.send_msg(Message::Ping(Vec::new())).await
     }
 
     async fn close(mut self) -> Result<()> {