+++ testgen/src/apalache.rs
@@ -88,6 +88,8 @@ pub fn run_apalache_test(dir: &str, test: ApalacheTestCase) -> io::Result<Apalac
     }
     cmd.arg("check");
     cmd.arg_from_parts(vec!["--inv=", &inv]);
+    cmd.arg("--init=InitTest");
+    cmd.arg("--next=NextTest");
     if let Some(length) = test.length {
         cmd.arg_from_parts(vec!["--length=", &length.to_string()]);
     }