+++ testgen/src/tester.rs
@@ -35,6 +35,10 @@ impl TestEnv {
         &self.current_dir
     }
 
+    pub fn clear_log(&self) -> Option<()> {
+        fs::remove_file(self.full_path("log")).ok()
+    }
+
     pub fn logln(&self, msg: &str) -> Option<()> {
         println!("{}", msg);
         fs::OpenOptions::new()
@@ -86,6 +90,11 @@ impl TestEnv {
     /// Returns None if copying was not successful
     pub fn copy_file_from_env(&self, other: &TestEnv, path: impl AsRef<Path>) -> Option<()> {
         self.copy_file_from(other.full_path(path))
+    }
+
+    /// Remove a file from a path relative to the environment current dir
+    pub fn remove_file(&self, rel_path: impl AsRef<Path>) -> Option<()> {
+        fs::remove_file(self.full_path(rel_path)).ok()
     }
 
     /// Convert a relative path to the full path from the test root
@@ -202,9 +211,7 @@ impl Tester {
 
     pub fn output_env(&self) -> Option<TestEnv> {
         let output_dir = self.root_dir.clone() + "/_" + &self.name;
-        fs::create_dir_all(&output_dir)
-            .ok()
-            .and(TestEnv::new(&output_dir))
+        TestEnv::new(&output_dir)
     }
 
     fn capture_test<F>(test: F) -> TestResult
@@ -270,8 +277,8 @@ impl Tester {
                 let env = TestEnv::new(dir.path().to_str().unwrap()).unwrap();
                 let output_dir = output_env.full_path(path);
                 let output_env = TestEnv::new(output_dir.to_str().unwrap()).unwrap();
-                fs::remove_dir_all(&output_dir).unwrap();
                 test(test_case, &env, &test_env, &output_env);
+                fs::remove_dir_all(&env.current_dir()).unwrap();
             }),
             Err(_) => ParseError,
         };